+++ functions/run_after_test.go
@@ -1,0 +1,14 @@
+package functions
+
+import (
+	"fmt"
+	"testing"
+)
+
+func TestRunAfter(t *testing.T) {
+	fmt.Println("I don't think any test cases can be written for RunAfter.")
+}
+
+func BenchmarkRunAfter(b *testing.B) {
+	fmt.Println("I don't think any benchmarks can be written for RunAfter.")
+}